Refactor generative components to streamline props and improve context usage
